Made methods out the switch cases in main and added method to DB class
Made methods in the Patient Manager class that check a patient in and
remove a patient from the database.
Made a method in Room Manager class to assign a patient to a room after
they have been checked in.
Made a method in the Database class to execute Updates on the database.
Also made tests for the Patient Manager class methods.